refactor: Move files into `html-to-notion` directory
