+++ src/content/sync/html-to-notion/__tests__/__fixtures__/formatting.html
@@ -18,6 +18,10 @@ monospace <strong>bold</strong>?
 
 final line</pre
   >
+  <blockquote>
+    <p>quotation start</p>
+    <p>quotation end</p>
+  </blockquote>
   <p>
     <span class="citation">()</span>
   </p>
@@ -25,12 +29,20 @@ final line</pre
   <p></p>
   <ul>
     <li><span style="color: #ff2020">red</span> one</li>
-    <li><span style="color: #ff7700">orange</span> two</li>
-    <li><span style="color: #ffcb00">yellow</span> three</li>
-    <li><span style="color: #4eb31c">green</span> four</li>
-    <li><span style="color: #05a2ef">blue</span> five</li>
-    <li><span style="color: #7953e3">purple</span> six</li>
-    <li><span style="color: #eb52f7">pink</span> seven</li>
+    <li><span style="color: #ff7700">orange</span> two<br />additional line</li>
+    <li>
+      <p><span style="color: #ffcb00">yellow</span> three</p>
+      <ul>
+        <li><span style="color: #4eb31c">green</span> four</li>
+        <li>
+          <p><span style="color: #05a2ef">blue</span> five</p>
+          <ul>
+            <li><span style="color: #eb52f7">pink</span> six</li>
+          </ul>
+        </li>
+        <li><span style="color: #7953e3">purple</span> seven</li>
+      </ul>
+    </li>
     <li><span style="color: #7e8386">gray</span> eight</li>
   </ul>
   <p></p>
