+++ src/content/sync/html-to-notion/__tests__/__fixtures__/index.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 
-import type { ChildBlock } from '../../notion-types';
+import type { ChildBlock } from '../../../notion-types';
 
 import * as formatting from './formatting';
 import * as simple from './simple';
@@ -17,7 +17,7 @@ const cases: Record<string, Pick<NoteTestCase, 'expected'>> = {
   formatting,
 };
 
-export const noteTestCases: NoteTestCase[] = Object.entries(cases).map(
+export const htmlTestCases: NoteTestCase[] = Object.entries(cases).map(
   ([name, { expected }]) => ({
     name,
     html: fs.readFileSync(path.resolve(__dirname, `${name}.html`), 'utf8'),
