+++ src/content/sync/html-to-notion/__tests__/__fixtures__/simple.ts
@@ -1,4 +1,4 @@
-import type { ChildBlock } from '../../notion-types';
+import type { ChildBlock } from '../../../notion-types';
 
 export const expected: ChildBlock[] = [
   {
