+++ src/content/sync/html-to-notion/html-to-notion.ts
@@ -1,20 +1,26 @@
 import 'core-js/stable/string/trim-end';
 import 'core-js/stable/string/trim-start';
 
-import { chunkString, getDOMParser } from '../utils';
+import { chunkString } from '../../utils';
 
 import type {
   Annotations,
   ChildBlock,
   ChildBlockType,
-  Color,
   ParagraphBlock,
   RichText,
   RichTextText,
   TextLink,
-} from './notion-types';
-
-type HTMLElementTagName = Uppercase<keyof HTMLElementTagNameMap>;
+} from '../notion-types';
+import { getAnnotations, getNotionColor } from './annotations';
+import {
+  getRootElement,
+  HTMLElementTagName,
+  isHTMLAnchorElement,
+  isHTMLBRElement,
+  isHTMLElement,
+  isTextNode,
+} from './dom-utils';
 
 // https://developer.mozilla.org/en-US/docs/Web/HTML/Block-level_elements#elements
 const TAG_BLOCK_TYPES = {
@@ -40,109 +46,8 @@ type SupportedBlock = Extract<ChildBlock, { type?: SupportedBlockType }>;
 
 type BlockElement = HTMLElement & { tagName: SupportedTagName };
 
-const TAG_ANNOTATIONS = {
-  B: { bold: true },
-  CODE: { code: true },
-  EM: { italic: true },
-  I: { italic: true },
-  S: { strikethrough: true },
-  STRONG: { bold: true },
-  U: { underline: true },
-} satisfies Partial<Record<HTMLElementTagName, Annotations>>;
-
-function canAnnotateTag(
-  tagName: string
-): tagName is keyof typeof TAG_ANNOTATIONS {
-  return tagName in TAG_ANNOTATIONS;
-}
-
-// https://www.zotero.org/support/note_templates#variables
-const ZOTERO_COLORS: Record<string, Color> = {
-  '255, 32, 32': 'red', // #ff2020
-  '255, 119, 0': 'orange', // #ff7700
-  '255, 203, 0': 'yellow', // #ffcb00
-  '78, 179, 28': 'green', // #4eb31c
-  '5, 162, 239': 'blue', // #05a2ef
-  '121, 83, 227': 'purple', // #7953e3
-  '235, 82, 247': 'pink', // #eb52f7
-  '126, 131, 134': 'gray', // #7e8386
-  '255, 102, 102': 'red_background', // #ff6666
-  '241, 152, 55': 'orange_background', // #f19837
-  '255, 212, 0': 'yellow_background', // #ffd400
-  '95, 178, 54': 'green_background', // #5fb236
-  '46, 168, 229': 'blue_background', // #2ea8e5
-  '162, 138, 229': 'purple_background', // #a28ae5
-  '229, 110, 238': 'pink_background', // #e56eee
-  '170, 170, 170': 'gray_background', // #aaaaaa
-};
-
-function getNotionColorFromString(color: string): Color {
-  if (!color) return;
-
-  const matches = color.match(/^rgba?\((\d+,\s*\d+,\s*\d+)/);
-  const rgb = matches?.[1];
-
-  return (rgb && ZOTERO_COLORS[rgb]) || undefined;
-}
-
-function getNotionColor(element: HTMLElement): Color {
-  return (
-    getNotionColorFromString(element.style.backgroundColor) ||
-    getNotionColorFromString(element.style.color)
-  );
-}
-
-function getAnnotations(element: HTMLElement): Annotations {
-  const annotations: Annotations = {};
-
-  const color = getNotionColor(element);
-
-  if (color) {
-    Object.assign(annotations, { color });
-  }
-
-  if (element.style.textDecoration === 'line-through') {
-    Object.assign(annotations, { strikethrough: true });
-  }
-
-  if (canAnnotateTag(element.tagName)) {
-    Object.assign(annotations, TAG_ANNOTATIONS[element.tagName]);
-  }
-
-  return Object.keys(annotations).length ? annotations : undefined;
-}
-
-/** https://developer.mozilla.org/en-US/docs/Web/API/Node/nodeType */
-enum NodeType {
-  ELEMENT_NODE = 1,
-  ATTRIBUTE_NODE = 2,
-  TEXT_NODE = 3,
-  CDATA_SECTION_NODE = 4,
-  PROCESSING_INSTRUCTION_NODE = 7,
-  COMMENT_NODE = 8,
-  DOCUMENT_NODE = 9,
-  DOCUMENT_TYPE_NODE = 10,
-  DOCUMENT_FRAGMENT_NODE = 11,
-}
-
-function isTextNode(node: Node): boolean {
-  return node.nodeType === NodeType.TEXT_NODE;
-}
-
 function isBlockElement(node: Node): node is BlockElement {
   return node.nodeName in TAG_BLOCK_TYPES;
-}
-
-function isHTMLElement(node: Node): node is HTMLElement {
-  return node.nodeType === NodeType.ELEMENT_NODE;
-}
-
-function isHTMLAnchorElement(node: Node): node is HTMLAnchorElement {
-  return node.nodeName === 'A';
-}
-
-function isHTMLBRElement(node: Node): node is HTMLBRElement {
-  return node.nodeName === 'BR';
 }
 
 function getMathExpression(element: Element): string | undefined {
@@ -157,12 +62,6 @@ function getMathExpression(element: Element): string | undefined {
 
 // https://developers.notion.com/reference/request-limits#limits-for-property-values
 const TEXT_CONTENT_MAX_LENGTH = 2000;
-
-function getRootElement(htmlString: string): Element | null {
-  const domParser = getDOMParser();
-  const doc = domParser.parseFromString(htmlString, 'text/html');
-  return doc.querySelector('body > div[data-schema-version]');
-}
 
 function buildBlock(element: Element): ChildBlock {
   // if (!element.hasChildNodes()) {
@@ -336,7 +235,7 @@ function collapseWhitespace(text: string): string {
   return text.replace(/[\s\n]+/g, ' ');
 }
 
-export function buildNoteBlocks(htmlString: string): ChildBlock[] {
+export function convertHtmlToBlocks(htmlString: string): ChildBlock[] {
   const root = getRootElement(htmlString);
   if (!root) throw new Error('Failed to load note content');
