+++ src/content/sync/html-to-notion/index.ts
@@ -1,0 +1,1 @@
+export { convertHtmlToBlocks } from './html-to-notion';
